refactor(playground): extract small UI features
Extract three independent features from main.ts:

features/phase-strip/ — phase label + progress bar updates
features/keyboard-shortcuts/ — shortcut sheet toggle
features/scoreboard/ — metric rows, sparklines, verdict ribbon,
  diff logic (MetricKey, METRIC_KEYS, diffMetrics, verdictToWinner)

Each feature defines narrow interface types for its state/UI deps
rather than importing the full State/UiHandles, keeping features
decoupled from the shell.
