+++ playground/src/features/scoreboard/index.ts
@@ -1,0 +1,11 @@
+export {
+  type MetricKey,
+  type Verdict,
+  type MetricVerdicts,
+  METRIC_KEYS,
+  METRIC_HISTORY_LEN,
+  diffMetrics,
+  initMetricRows,
+  renderMetricRows,
+} from "./metric-rows";
+export { renderVerdictRibbon } from "./verdict-ribbon";
